Organized topics components into subfolders.
